requirements.txt
httpx==0.28.1
langchain_openai==0.3.0
Pillow==11.1.0
streamlit>=1.41.1
